requirements.txt
# Core
google-genai>=1.0.0
Pillow>=11.0.0
python-dotenv>=1.0.0

# Ingest
PyMuPDF>=1.25.0

# Server
fastapi>=0.115.0
uvicorn>=0.34.0
watchfiles>=1.0.0

# Voice proxy
httpx>=0.28.0
